+++ packages/salesforcedx-vscode/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 52.4.0 - July 8, 2021
+
+## Added
+
+#### salesforcedx-vscode-core
+
+- Enable conflict detection for all deploy commands ([PR #3357](https://github.com/forcedotcom/salesforcedx-vscode/pull/3357))
+
 # 52.3.0 - June 30, 2021
 
 ## Added